--- a/WFInfo/Ocr.cs
+++ b/WFInfo/Ocr.cs
@@ -1638,7 +1638,7 @@ namespace WFInfo
             }
             else
             {
-                partialScreenshotExpanded = image;
+                partialScreenshotExpanded = (Bitmap)image.Clone();
                 filtered = image;
             }
             Color clr;
@@ -1646,7 +1646,7 @@ namespace WFInfo
             int numbytes = Math.Abs(lockedBitmapData.Stride) * lockedBitmapData.Height;
             byte[] LockedBitmapBytes = new byte[numbytes];
             Marshal.Copy(lockedBitmapData.Scan0, LockedBitmapBytes, 0, numbytes);
-            int PixelSize = 3; //24 bits/8 = 3 bytes
+            int PixelSize = 4; //24 bits/8 = 3 bytes
             for (int x = 0; x < filtered.Width; x++)
             {
                 for (int y = 0; y < filtered.Height; y++)
@@ -1657,17 +1657,20 @@ namespace WFInfo
                         LockedBitmapBytes[(y * filtered.Width + x) * PixelSize] = 0;
                         LockedBitmapBytes[(y * filtered.Width + x) * PixelSize + 1] = 0;
                         LockedBitmapBytes[(y * filtered.Width + x) * PixelSize + 2] = 0;
+                        LockedBitmapBytes[(y * filtered.Width + x) * PixelSize + 3] = 0;
                         //filtered.SetPixel(x, y, Color.Black);
                     } else
                     {
                         LockedBitmapBytes[(y * filtered.Width + x) * PixelSize] = 255;
                         LockedBitmapBytes[(y * filtered.Width + x) * PixelSize + 1] = 255;
                         LockedBitmapBytes[(y * filtered.Width + x) * PixelSize + 2] = 255;
+                        LockedBitmapBytes[(y * filtered.Width + x) * PixelSize + 3] = 255;
                         //filtered.SetPixel(x, y, Color.White);
                     }
                 }
             }
             Marshal.Copy(LockedBitmapBytes, 0, lockedBitmapData.Scan0, numbytes);
+            filtered.UnlockBits(lockedBitmapData);
             return filtered;
         }