### What did you fix? (provide a description or issue closes statement)
Sped up pixel get and set times for filtering step of Snap It and most of Master It, as suggested by CruleD#8893 on discord ([reference implementation](https://github.com/CruleD/Warframe-ReBa-Farmer/blob/2745fed6a4847659608e19a144fe71b86d256c02/Warframe%20ReBa%20Farmer/Module_RelicScanner.cs#L170))

---

### Reproduction steps
1. Use Snap It or Master It. Debug mode and a few runs per test image recommended to get a good time comparison

---

### Evidence/screenshot/link to line
For the following test image, time went from around 2500ms to around 650 ms. (snap it, whole image due to debug mode, count item disabled)
![FullScreenShot ](https://user-images.githubusercontent.com/77355844/129268260-b6a2dbba-5dc6-4443-ad4c-452e085caff6.png)

For the following test image, time went from around 1600ms to around 1400ms (mastered scan)
![ProfileImage 2021-08-12 15-38-3490](https://user-images.githubusercontent.com/77355844/129268626-bb16a74e-99f0-4b15-8bb9-8ffbb513c2de.png)

For the following test image, time went from **around 12700ms (!!!)** to around 1500ms (mastered scan)
![HighWhiteCloud](https://user-images.githubusercontent.com/77355844/129269018-cc1f3db2-3db0-4dda-b739-296315f91f2e.png)


### Considerations
- Does this contain a new dependency? **[Yes]** (System.Drawing.Imaging)
- Does this introduce opinionated data formatting or manual data entry? **[No]**
- Does this pr include updated data files in a separate commit that can be reverted for a clean code-only pr? **[No]**
- Have I run the linter? **[No]**
- Is is a bug fix, feature request, or enhancement? **[Enhancement]**